pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentinel-log-ai-ml"
version = "0.1.0"
description = "AI/ML engine for sentinel-log-ai - embeddings, clustering, novelty detection, and LLM explanations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Sentinel Log AI Team" }
]
keywords = ["logging", "ai", "ml", "clustering", "llm", "observability", "embeddings", "grpc"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "structlog>=23.1",
    "grpcio>=1.59",
    "grpcio-tools>=1.59",
    "protobuf>=4.24",
]

[project.optional-dependencies]
ml = [
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7",
    "hdbscan>=0.8",
    "scikit-learn>=1.3",
    "torch>=2.0",
]
llm = [
    "ollama>=0.1",
    "httpx>=0.25",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "mypy>=1.5",
    "ruff>=0.1",
    "pre-commit>=3.4",
    "types-PyYAML>=6.0",
    "grpcio-testing>=1.59",
    "mypy-protobuf>=3.5",
]
all = [
    "sentinel-log-ai-ml[ml,llm,dev]",
]

[project.scripts]
sentinel-ml = "sentinel_ml.server:main"

[tool.hatch.build.targets.wheel]
packages = ["python/sentinel_ml"]

# ============================================================================
# Ruff Configuration
# ============================================================================
[tool.ruff]
target-version = "py310"
line-length = 100
src = ["python", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM", "TCH", "PTH", "RUF"]
ignore = ["E501", "B008", "B904"]

[tool.ruff.lint.isort]
known-first-party = ["sentinel_ml"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["ARG001", "S101"]
"python/sentinel_ml/generated/**/*" = ["ALL"]
"python/sentinel_ml/logging.py" = ["TC003"]  # MutableMapping used at runtime in function signatures

# ============================================================================
# Mypy Configuration
# ============================================================================
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = ["faiss.*", "hdbscan.*", "sentence_transformers.*", "ollama.*", "grpc.*", "google.protobuf.*"]
ignore_missing_imports = true

# ============================================================================
# Pytest Configuration
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests/python"]
python_files = ["test_*.py"]
addopts = ["-ra", "-q", "--strict-markers", "-v"]
markers = [
    "slow: marks tests as slow",
    "integration: integration tests",
    "ml: requires ML dependencies",
    "llm: requires LLM dependencies",
]

[tool.coverage.run]
source = ["python/sentinel_ml"]
branch = true
